requirements.txt
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
et_xmlfile==2.0.0
joblib==1.5.1
loguru==0.7.3
nltk==3.9.1
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2024.11.6
scikit-learn==1.7.1
scipy==1.16.0
six==1.17.0
SQLAlchemy==2.0.41
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
